requirements.txt
alpaca-py==0.43.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
# Editable install with no version control (momentum-signal-engine==0.1.0)
-e /Users/nayeem.mohammad/projects/momentum-signal-engine
msgpack==1.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rich==14.3.2
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sseclient-py==1.9.0
starlette==0.52.1
ta==0.11.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
